# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ramon api",
        "description": "api for games, movies and books",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/game/{id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get game by index",
                "description": "Returns a game by the given id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Edit game",
                "description": "Updates a game Consuming a Game object",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Game object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "description": "Delete a game with the given id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/game": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Creates and returns a game object",
                "description": "Creates and returns a game",
                "operationId": "create",
                "requestBody": {
                    "description": "Create a Game object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/game/giveTag/{id}": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "add a tag to a game",
                "description": "add tag",
                "operationId": "tagGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}